LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/multi-globals/workers-coep-report.https-expected.txt

CONSOLE MESSAGE: Refused to load 'https://localhost:9443/html/cross-origin-embedder-policy/multi-globals/current/worker.js' worker because of Cross-Origin-Embedder-Policy.
CONSOLE MESSAGE: Worker load was blocked by Cross-Origin-Embedder-Policy
CONSOLE MESSAGE: Cannot load https://localhost:9443/html/cross-origin-embedder-policy/multi-globals/current/worker.js due to access control checks.
 Hello

FAIL Multiple globals for Worker constructor: COEP reports assert_equals: expected 1 but got 0
